Added flightpath csv reader
