+++ flights/demo_flight.csv
@@ -1,0 +1,5 @@
+segment, startxy, endxy, speeds
+constant-speed flyby, -200 30, 180 30, 30 30
+decelerate, 180 30, 200 30, 30 0
+accelerate, 200 30, 180 30, 0 30
+flyby back, 180 30, -200 30, 30 30
