+++ src/utils.py
@@ -1,0 +1,28 @@
+import numpy as np
+
+
+def test_sine(t, f_0=440, A=0.75, fs=48000):
+    t_samples = np.linspace(0, t, fs*t)
+    sig = np.zeros_like(t_samples)
+
+    for i, sample in enumerate(t_samples):
+        sig[i] = np.sin(2*np.pi*f_0*sample) * A
+
+    return sig
+
+
+def load_params(csv_file):
+    str_params = np.loadtxt(
+        csv_file,
+        delimiter=',',
+        skiprows=1,
+        usecols=np.arange(1, 4),
+        dtype='str'
+    )
+
+    params = np.zeros((*str_params.shape, 2))
+    for i, row in enumerate(str_params):
+        for j, cell in enumerate(row):
+            params[i, j, :] = np.array(cell.strip().split(' '), dtype=int)
+
+    return params
